cardapio_ai/requirements.txt
flask>=3.0.0,<4.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
openai>=1.0.0
groq>=0.11.0